llm-sophia/scripts/train.py
#!/usr/bin/env python3
"""
Sophia LLM Training Script
Run with: python scripts/train.py
"""

import os
import sys
import yaml
from pathlib import Path

# Add src to path
sys.path.append(str(Path(__file__).parent.parent / "src"))

def main():
    print("🧠 Sophia LLM Training")
    print("This script will be implemented as the training pipeline develops.")
    print("For now, please follow the implementation roadmap.")
    
    # Load configuration
    config_path = Path(__file__).parent.parent / "configs" / "sophia_config.yaml"
    if config_path.exists():
        with open(config_path) as f:
            config = yaml.safe_load(f)
        print(f"✅ Configuration loaded: {config['project_name']}")
    else:
        print("❌ Configuration file not found")

if __name__ == "__main__":
    main()
